+++ ecs/cache.go
@@ -5,6 +5,7 @@ type cacheEntry struct {
 	ID         uint32              // Filter ID.
 	Filter     Filter              // The underlying filter.
 	Archetypes pointers[archetype] // Nodes matching the filter.
+	Indices    map[*archetype]int  // Map of archetype indices for removal.
 }
 
 // Cache provides [Filter] caching to speed up queries.
@@ -54,6 +55,7 @@ func (c *Cache) Register(f Filter) CachedFilter {
 			ID:         id,
 			Filter:     f,
 			Archetypes: pointers[archetype]{c.getArchetypes(f)},
+			Indices:    nil,
 		})
 	c.indices[id] = len(c.filters) - 1
 	return CachedFilter{f, id}
@@ -114,29 +116,46 @@ func (c *Cache) addArchetype(arch *archetype) {
 		if rf, ok := e.Filter.(*RelationFilter); ok {
 			if rf.Target == arch.RelationTarget {
 				e.Archetypes.Add(arch)
+				// Not required: can't add after removing,
+				// as the target entity is dead.
+				// if e.Indices != nil { e.Indices[arch] = int(e.Archetypes.Len() - 1) }
 			}
 			continue
 		}
 		e.Archetypes.Add(arch)
+		if e.Indices != nil {
+			e.Indices[arch] = int(e.Archetypes.Len() - 1)
+		}
 	}
 }
 
 // Removes an archetype.
 //
-// Iterates over all filters and removes the archetype from the resp. entry where the filter matches.
+// Can only be used for archetypes that have a relation target.
+// Archetypes without a relation are never removed.
 func (c *Cache) removeArchetype(arch *archetype) {
 	for i := range c.filters {
 		e := &c.filters[i]
-		if !e.Filter.Matches(arch.Mask) {
-			continue
+
+		if e.Indices == nil && e.Filter.Matches(arch.Mask) {
+			c.mapArchetypes(e)
 		}
 
-		if rf, ok := e.Filter.(*RelationFilter); ok {
-			if rf.Target == arch.RelationTarget {
-				e.Archetypes.Remove(arch)
+		if idx, ok := e.Indices[arch]; ok {
+			swap := e.Archetypes.RemoveAt(idx)
+			if swap {
+				e.Indices[e.Archetypes.Get(int32(idx))] = idx
 			}
-			continue
+			delete(e.Indices, arch)
 		}
-		e.Archetypes.Remove(arch)
+	}
+}
+
+func (c *Cache) mapArchetypes(e *cacheEntry) {
+	e.Indices = map[*archetype]int{}
+	for i, arch := range e.Archetypes.pointers {
+		if arch.HasRelation() {
+			e.Indices[arch] = i
+		}
 	}
 }