+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 * Speed up archetype and node iteration to be as fast as before the new nested structure (#270, #288)
 * ~~Filter cache stores archetype graph nodes instead of archetypes (#276)~~ (#288)
 * Use `uint32` instead of `uintptr` for indices and query iteration counter (#283)
+* Cached filters use a map for faster removal of archetypes (#289)
 
 ### Documentation